EO-894 - Rename components
