+++ src/pages/alerts/tests/EditPage.test.js
@@ -2,7 +2,7 @@ import { shallow } from 'enzyme';
 import React from 'react';
 import { EditPage } from '../EditPage';
 import formatFormValues from '../helpers/formatFormValues';
-import AlertFormNew from '../components/AlertFormNew';
+import AlertForm from '../components/AlertForm';
 
 jest.mock('../helpers/formatFormValues');
 
@@ -49,7 +49,7 @@ describe('Page: Alerts Edit', () => {
 
   it('should handle submit', async () => {
     formatFormValues.mockImplementationOnce((a) => a);
-    await wrapper.find(AlertFormNew).simulate('submit', { value: 'mock value' });
+    await wrapper.find(AlertForm).simulate('submit', { value: 'mock value' });
     expect(props.updateAlert).toHaveBeenCalledWith({ data: { value: 'mock value' }, id: props.id });
     expect(props.showUIAlert).toHaveBeenCalled();
     expect(props.history.push).toHaveBeenCalledWith('/alerts/details/alert-id-1');
